Move count_action out of player.cc
It has no particular relation to player functions, and it makes
testing changes a real PITA.
